requirements.txt
requests
pybliometrics
python-dotenv
pandas
matplotlib
flask 